requirements.txt
amqp==5.0.1
asgiref==3.2.7
bcrypt==3.1.7
billiard==3.6.3.0
celery==5.0.0
certifi==2020.11.8
cffi==1.14.0
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
Django==3.1
djangorestframework==3.11.0
elasticsearch==5.5.3
idna==2.10
kombu==5.0.2
limits==1.5.1
mysqlclient==1.4.6
prompt-toolkit==3.0.7
pycparser==2.20
pylogbeat==2.0.0
python-logstash-async==2.1.1
pytz==2020.1
requests==2.25.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
